[GVMWin-beta] BetaV1 --> +Scan +all others functionalities done!

+++ cmd/switch.go
@@ -5,6 +5,7 @@ Copyright © 2022 Ilingu <ilingu@protonmail.com>
 package cmd
 
 import (
+	"errors"
 	"fmt"
 	"gvm-windows/cmd/utils"
 	"gvm-windows/gvm"
@@ -21,7 +22,7 @@ import (
 var switchCmd = &cobra.Command{
 	Use:   "switch",
 	Short: "➡️ Let you switch of Go version easily (recommended).",
-	Long:  "Let you switch of Go Main Version easily: uses the Go MSI executable (recommended for windows).\nIt Downloads the msi executable from Go Official Website if not already downloaded, then it uninstalls the current Go version and finally it installs the newly/already downloaded Go version.\nTIP: use 'latest' arg to download latest go version",
+	Long:  "➡️ Let you switch of Go Main Version easily: uses the Go MSI executable (recommended for windows).\nIt Downloads the msi executable from Go Official Website if not already downloaded, then it uninstalls the current Go version and finally it installs the newly/already downloaded Go version.\nTIP: use 'latest' arg to download latest go version",
 	Run: func(cmd *cobra.Command, args []string) {
 		if !utils.IsArgsValids(args) {
 			cmd.Help()
@@ -30,14 +31,14 @@ var switchCmd = &cobra.Command{
 		version := args[0]
 
 		latestVersion, ok := utils.GetLatestGoVersion()
-		if !ok {
+		if version == "latest" && !ok {
 			log.Println("❌ Cannot fetch latest Go Version. Check your internet connection!")
 			return
 		}
 
 		if version == "latest" {
 			version = latestVersion
-		} else {
+		} else if runtime.Version() != "go"+latestVersion {
 			log.Printf("❕ Go's latest version is %s, you're currently in %s", latestVersion, runtime.Version())
 		}
 
@@ -55,29 +56,35 @@ var switchCmd = &cobra.Command{
 
 		GoMsiExecutable, ok := appFolder+fmt.Sprintf("\\go%s.msi", version), false
 		Godl := gvm.MakeGoDownloader(version)
-		if temp {
-			log.Printf("Downloading go%s... ⏳\n", version)
+		downloadGo := func() error {
+			log.Printf("Downloading go%s... ⏳ (no-cache=%t)\n", version, temp)
 
-			GoMsiExecutable, ok = Godl.DownloadTempMSI()
-			defer os.Remove(GoMsiExecutable) // Remove Temp File
+			if temp {
+				GoMsiExecutable, ok = Godl.DownloadTempMSI()
+			} else {
+				GoMsiExecutable, ok = Godl.DownloadMSI()
+			}
+
 			if !ok {
 				log.Printf("Failed to download go%s ❌\n", version)
+				return errors.New("")
+			}
+
+			log.Printf("go%s Downloaded Successfully ✅: %s\n", version, GoMsiExecutable)
+			return nil
+		}
+
+		fileInfo, err := os.Stat(GoMsiExecutable)
+		if !temp && (os.IsNotExist(err) || err != nil || fileInfo.IsDir()) {
+			log.Println("❌ This Go Version is not downloaded on your machine!")
+			if downloadGo() != nil {
 				return
 			}
-			log.Printf("go%s Downloaded Successfully ✅: %s\n", version, GoMsiExecutable)
-		} else {
-			fileInfo, err := os.Stat(GoMsiExecutable)
-			if !temp && (os.IsNotExist(err) || err != nil || fileInfo.IsDir()) {
-				log.Println("❌ This Go Version is not downloaded on your machine!")
-				log.Printf("Downloading go%s... ⏳\n", version)
-
-				GoMsiExecutable, ok = Godl.DownloadMSI()
-				if !ok {
-					log.Printf("Failed to download go%s ❌\n", version)
-					return
-				}
-				log.Printf("go%s Downloaded Successfully ✅: %s\n", version, GoMsiExecutable)
+		} else if temp {
+			if downloadGo() != nil {
+				return
 			}
+			defer os.Remove(GoMsiExecutable) // Remove Temp File
 		}
 
 		log.Printf("Installing go%s... ⏳\n", version)
